+++ packages/terra-form-textarea/src/Textarea.jsx
@@ -66,6 +66,10 @@ const propTypes = {
    */
   onFocus: PropTypes.func,
   /**
+   * Placeholder text.
+   */
+  placeholder: PropTypes.string,
+  /**
    * Whether the input is required or not.
    */
   required: PropTypes.bool,
@@ -97,6 +101,7 @@ const defaultProps = {
   isIncomplete: false,
   isInvalid: false,
   onChange: undefined,
+  placeholder: undefined,
   required: false,
   rows: null,
   size: 'small',
@@ -185,6 +190,7 @@ class Textarea extends React.Component {
       required,
       onChange,
       onFocus,
+      placeholder,
       isAutoResizable,
       isIncomplete,
       isInvalid,
@@ -244,6 +250,7 @@ class Textarea extends React.Component {
         name={name}
         onFocus={this.onFocus}
         onChange={this.onChange}
+        placeholder={placeholder}
         required={required}
         rows={textareaRows}
         className={textareaClasses}