+++ packages/terra-form-textarea/src/TextareaField.jsx
@@ -76,6 +76,10 @@ const propTypes = {
    */
   onChange: PropTypes.func,
   /**
+   * Placeholder text.
+   */
+  placeholder: PropTypes.string,
+  /**
    * Whether or not the field is required.
    */
   required: PropTypes.bool,
@@ -107,6 +111,7 @@ const defaultProps = {
   labelAttrs: {},
   maxWidth: undefined,
   onChange: undefined,
+  placeholder: undefined,
   required: false,
   showOptional: false,
   value: undefined,
@@ -132,6 +137,7 @@ const TextareaField = (props) => {
     required,
     showOptional,
     onChange,
+    placeholder,
     value,
     ...customProps
   } = props;
@@ -173,6 +179,7 @@ const TextareaField = (props) => {
         id={inputId}
         isIncomplete={isIncomplete}
         onChange={onChange}
+        placeholder={placeholder || inputAttrs.placeholder}
         value={value}
         defaultValue={defaultValue}
         aria-describedby={ariaDescriptionIds}